part-4 complete

+++ controllers/blog.js
@@ -1,8 +1,13 @@
 const blogRouter = require("express").Router();
+const { userExtractor } = require("../utils/middleware");
 const Blog = require("../models/blog");
+const User = require("../models/users");
 
 blogRouter.get("/", async (req, res) => {
-  const data = await Blog.find({});
+  const data = await Blog.find({}).populate("user", {
+    password: 0,
+    notes: 0,
+  });
   res.json(data);
 });
 
@@ -15,20 +20,31 @@ blogRouter.get("/:id", async (req, res, next) => {
   }
 });
 
-blogRouter.post("/", async (req, res, next) => {
+blogRouter.post("/", userExtractor, async (req, res, next) => {
   try {
-    const blog = new Blog(req.body);
+    const user = await User.findById(req.user);
+    const blog = new Blog({ ...req.body, user: user._id });
+
     const data = await blog.save();
+    user.notes = [...user.notes, data._id];
+
+    user.save();
     res.json(data);
   } catch (err) {
     next(err);
   }
 });
 
-blogRouter.delete("/:id", async (req, res, next) => {
+blogRouter.delete("/:id", userExtractor, async (req, res, next) => {
   try {
-    await Blog.findByIdAndDelete(req.params.id);
-    res.status(200).end();
+    const blog = await Blog.findById(req.params.id);
+
+    if (blog && blog.user.toString() === req.user.toString()) {
+      await Blog.findByIdAndDelete(req.params.id);
+      res.status(200).end();
+    } else {
+      res.status(401).end();
+    }
   } catch (err) {
     next(err);
   }
